requirements.txt
matplotlib==3.10.7
numpy==2.3.4
pandas==2.3.3
PyYAML==6.0.3
scikit_learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
xgboost==3.1.1
